requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
WTForms==3.0.1
PyJWT==2.8.0
redis==5.0.1
celery==5.3.4
pydantic==2.4.2
numpy==1.26.0
monty==2023.9.25
